working on the about page
