+++ src/components/layout/Hero1.jsx
@@ -1,0 +1,34 @@
+import React from 'react'
+// import '../../../index.css'
+
+
+function Hero1(){
+
+    
+        return (
+          <div>
+            <div className=''>
+          <div className="Hero-background">
+            
+            <div className=' grid items-center  justify-center '>
+            <h1 className="  text-5xl pt-44 ">About Us</h1>
+  
+         <a href='#' className='text-slate-200 pt-4 text-2xl'>
+         Home 
+         <i className="fa fa-angle-right pr-2"></i>
+         <span>About Us</span></a>
+
+            </div>
+           
+          </div>
+          
+
+          </div>
+          </div>
+        )
+    
+
+}
+
+
+export default Hero1
